+++ scandeval/benchmarks/abstract/ner.py
@@ -26,6 +26,9 @@ class NerBenchmark(TokenClassificationBenchmark):
         evaluate_train (bool, optional):
             Whether the models should be evaluated on the training scores.
             Defaults to False.
+        use_auth_token (bool, optional):
+            Whether the benchmark should use an authentication token. Defaults
+            to False.
         verbose (bool, optional):
             Whether to print additional output during evaluation. Defaults to
             False.
@@ -43,6 +46,7 @@ class NerBenchmark(TokenClassificationBenchmark):
         cache_dir (str): Directory where models are cached.
         two_labels (bool): Whether two labels should be predicted.
         split_point (int or None): Splitting point of `id2label` into labels.
+        use_auth_token (bool): Whether an authentication token should be used.
         verbose (bool): Whether to print additional output.
     '''
     def __init__(self,
@@ -50,6 +54,7 @@ class NerBenchmark(TokenClassificationBenchmark):
                  language: str,
                  cache_dir: str = '.benchmark_models',
                  evaluate_train: bool = False,
+                 use_auth_token: bool = False,
                  verbose: bool = False):
         id2label = ['B-LOC', 'I-LOC', 'B-ORG', 'I-ORG', 'B-PER',
                     'I-PER', 'B-MISC', 'I-MISC', 'O']
@@ -82,6 +87,7 @@ class NerBenchmark(TokenClassificationBenchmark):
                          label_synonyms=label_synonyms,
                          cache_dir=cache_dir,
                          evaluate_train=evaluate_train,
+                         use_auth_token=use_auth_token,
                          verbose=verbose)
 
     def _load_data(self) -> Tuple[Dataset, Dataset]: