+++ scandeval/benchmarks/abstract/pos.py
@@ -26,6 +26,9 @@ class PosBenchmark(TokenClassificationBenchmark):
         evaluate_train (bool, optional):
             Whether the models should be evaluated on the training scores.
             Defaults to False.
+        use_auth_token (bool, optional):
+            Whether the benchmark should use an authentication token. Defaults
+            to False.
         verbose (bool, optional):
             Whether to print additional output during evaluation. Defaults to
             False.
@@ -43,7 +46,7 @@ class PosBenchmark(TokenClassificationBenchmark):
         cache_dir (str): Directory where models are cached.
         two_labels (bool): Whether two labels should be predicted.
         split_point (int or None): Splitting point of `id2label` into labels.
-
+        use_auth_token (bool): Whether an authentication token should be used.
         verbose (bool): Whether to print additional output.
     '''
     def __init__(self,
@@ -51,6 +54,7 @@ class PosBenchmark(TokenClassificationBenchmark):
                  language: str,
                  cache_dir: str = '.benchmark_models',
                  evaluate_train: bool = False,
+                 use_auth_token: bool = False,
                  verbose: bool = False):
         id2label = ['ADJ', 'ADV', 'INTJ', 'NOUN', 'PROPN', 'VERB', 'ADP',
                     'AUX', 'CCONJ', 'DET', 'NUM', 'PART', 'PRON', 'SCONJ',
@@ -61,6 +65,7 @@ class PosBenchmark(TokenClassificationBenchmark):
                          id2label=id2label,
                          cache_dir=cache_dir,
                          evaluate_train=evaluate_train,
+                         use_auth_token=use_auth_token,
                          verbose=verbose)
 
     def _load_data(self) -> Tuple[Dataset, Dataset]: