+++ scandeval/benchmarks/ddt_pos.py
@@ -18,6 +18,9 @@ class DdtPosBenchmark(PosBenchmark):
         evaluate_train (bool, optional):
             Whether the models should be evaluated on the training scores.
             Defaults to False.
+        use_auth_token (bool, optional):
+            Whether the benchmark should use an authentication token. Defaults
+            to False.
         verbose (bool, optional):
             Whether to print additional output during evaluation. Defaults to
             False.
@@ -34,14 +37,17 @@ class DdtPosBenchmark(PosBenchmark):
         cache_dir (str): Directory where models are cached.
         two_labels (bool): Whether two labels should be predicted.
         split_point (int or None): Splitting point of `id2label` into labels.
+        use_auth_token (bool): Whether an authentication token should be used.
         verbose (bool): Whether to print additional output.
     '''
     def __init__(self,
                  cache_dir: str = '.benchmark_models',
                  evaluate_train: bool = False,
+                 use_auth_token: bool = False,
                  verbose: bool = False):
         super().__init__(name='ddt-pos',
                          language='da',
                          cache_dir=cache_dir,
                          evaluate_train=evaluate_train,
+                         use_auth_token=use_auth_token,
                          verbose=verbose)